openapi: Fix number -> integer issues

+++ client/model_cluster_profile_amazon_amazon_node_pools_pool1.go
@@ -14,8 +14,8 @@ type ClusterProfileAmazonAmazonNodePoolsPool1 struct {
 	InstanceType string `json:"instanceType,omitempty"`
 	SpotPrice string `json:"spotPrice,omitempty"`
 	Autoscaling bool `json:"autoscaling,omitempty"`
-	Count float32 `json:"count,omitempty"`
-	MinCount float32 `json:"minCount,omitempty"`
-	MaxCount float32 `json:"maxCount,omitempty"`
+	Count int32 `json:"count,omitempty"`
+	MinCount int32 `json:"minCount,omitempty"`
+	MaxCount int32 `json:"maxCount,omitempty"`
 	Image string `json:"image,omitempty"`
 }
